Add VM to Execution::cost
