+++ synthesizer/src/vm/execute.rs
@@ -35,21 +35,8 @@ impl<N: Network, C: ConsensusStorage<N>> VM<N, C> {
         let fee = match fee {
             None => None,
             Some((credits, priority_fee_in_microcredits)) => {
-                // Prepare the program lookup.
-                let lookup = execution
-                    .transitions()
-                    .map(|transition| {
-                        let program_id = transition.program_id();
-                        Ok((
-                            *program_id,
-                            self.transaction_store()
-                                .get_program(program_id)?
-                                .ok_or(anyhow!("Program {program_id} not found"))?,
-                        ))
-                    })
-                    .collect::<Result<HashMap<_, _>>>()?;
                 // Compute the minimum execution cost.
-                let (minimum_execution_cost, (_, _)) = Execution::cost(&execution, lookup)?;
+                let (minimum_execution_cost, (_, _)) = Execution::cost(self, &execution)?;
                 // Determine the fee.
                 let fee_in_microcredits = minimum_execution_cost
                     .checked_add(priority_fee_in_microcredits)
